requirement.txt
streamlit==1.48.1
pandas==2.2.3
google-generativeai==0.8.5
pillow==10.0
ollama==0.5.3